fix(src/helpers/axios.js): add conditions about the server error when the response have information

+++ src/helpers/axios.js
@@ -27,9 +27,9 @@ axiosInstance.interceptors.response.use(
       return res;
    },
    (error) => {
-      const { status } = error.response;
       console.log(error.response);
-      if (status === 500) {
+      const status = error.response ? error.response.status : 500;
+      if (status && status === 500) {
          localStorage.clear();
          store.dispatch({ type: authConstants.LOGOUT_SUCCESS });
       }
